requirements.txt
flask
flask-wtf
python-dotenv
flask-sqlalchemy
flask-mail
wtforms-sqlalchemy
pandas